+++ packages/shared/src/metadata-manager.ts
@@ -3,6 +3,7 @@ import {
   type MetadataEntry,
   type SettingsManager,
 } from "@saleor/app-sdk/settings-manager";
+import { type Logger } from "@saleor/apps-logger";
 import { type Client, gql } from "urql";
 
 import { createRotatingDecryptCallback } from "./key-rotation";
@@ -67,38 +68,72 @@ type UpdateAppPrivateMetadataMutation = {
 
 export type MetadataManagerGraphqlClient = Pick<Client, "mutation" | "query">;
 
-async function fetchAllPrivateMetadata(
-  client: MetadataManagerGraphqlClient,
-): Promise<MetadataEntry[]> {
+async function fetchAllPrivateMetadata({
+  client,
+  logger,
+}: {
+  client: MetadataManagerGraphqlClient;
+  logger: Logger;
+}): Promise<MetadataEntry[]> {
   const { error, data } = await client
     .query<FetchAppPrivateMetadataQuery>(FetchAppDetailsQuery, {})
     .toPromise();
 
   if (error) {
+    const cause =
+      error.networkError?.cause instanceof Error
+        ? (error.networkError.cause as NodeJS.ErrnoException)
+        : undefined;
+
+    logger.error("Failed to fetch app metadata", {
+      errorMessage: error.message,
+      networkErrorMessage: error.networkError?.message,
+      causeCode: cause?.code,
+      causeMessage: cause?.message,
+    });
+
     return [];
   }
 
   return data?.app?.privateMetadata.map((md) => ({ key: md.key, value: md.value })) || [];
 }
 
-async function updatePrivateMetadata(
-  client: MetadataManagerGraphqlClient,
-  metadata: MetadataEntry[],
-  appId: string,
-) {
-  const { error: mutationError, data: mutationData } = await client
+async function updatePrivateMetadata({
+  client,
+  metadata,
+  appId,
+  logger,
+}: {
+  client: MetadataManagerGraphqlClient;
+  metadata: MetadataEntry[];
+  appId: string;
+  logger: Logger;
+}) {
+  const { error, data } = await client
     .mutation<UpdateAppPrivateMetadataMutation>(UpdateAppMetadataMutation, {
       id: appId,
       input: metadata,
     })
     .toPromise();
 
-  if (mutationError) {
-    throw new Error(`Mutation error: ${mutationError.message}`);
+  if (error) {
+    const cause =
+      error.networkError?.cause instanceof Error
+        ? (error.networkError.cause as NodeJS.ErrnoException)
+        : undefined;
+
+    logger.error("Failed to update app metadata", {
+      errorMessage: error.message,
+      networkErrorMessage: error.networkError?.message,
+      causeCode: cause?.code,
+      causeMessage: cause?.message,
+    });
+
+    throw new Error(`Mutation error: ${error.message}`);
   }
 
   return (
-    mutationData?.updatePrivateMetadata?.item?.privateMetadata.map((md) => ({
+    data?.updatePrivateMetadata?.item?.privateMetadata.map((md) => ({
       key: md.key,
       value: md.value,
     })) || []
@@ -109,6 +144,7 @@ export class EncryptedMetadataManagerFactory {
   constructor(
     private encryptionKey: string,
     private fallbackKeys: string[] = [],
+    private logger: Logger,
   ) {
     if (!encryptionKey) {
       throw new Error("Encryption key is required");
@@ -118,9 +154,15 @@ export class EncryptedMetadataManagerFactory {
   create(client: MetadataManagerGraphqlClient, appId: string): SettingsManager {
     return new EncryptedMetadataManager({
       encryptionKey: this.encryptionKey,
-      fetchMetadata: () => fetchAllPrivateMetadata(client),
-      mutateMetadata: (metadata: MetadataEntry[]) => updatePrivateMetadata(client, metadata, appId),
-      async deleteMetadata(keys: string[]) {
+      fetchMetadata: () => fetchAllPrivateMetadata({ client, logger: this.logger }),
+      mutateMetadata: (metadata) =>
+        updatePrivateMetadata({
+          client,
+          metadata,
+          appId,
+          logger: this.logger,
+        }),
+      async deleteMetadata(keys) {
         await client.mutation(DeletePrivateMetadataMutation, {
           id: appId,
           keys: keys,