+++ apps/search/CHANGELOG.md
@@ -1,5 +1,13 @@
 # saleor-app-search
 
+## 1.28.1
+
+### Patch Changes
+
+- 622d13c: Updated GraphQL schema to 3.23
+- Updated dependencies [622d13c]
+  - @saleor/webhook-utils@0.2.10
+
 ## 1.28.0
 
 ### Minor Changes